@@ -6,43 +6,44 @@
     "start": "ng serve",
     "build": "ng build",
     "watch": "ng build --watch --configuration development",
-    "test": "ng test"
+    "test": "ng test",
+    "lint": "ng lint"
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "~13.2.6",
-    "@angular/common": "~13.2.6",
-    "@angular/compiler": "~13.2.6",
-    "@angular/core": "~13.2.6",
-    "@angular/forms": "^13.2.6",
-    "@angular/platform-browser": "~13.2.6",
-    "@angular/platform-browser-dynamic": "~13.2.6",
-    "@angular/router": "~13.2.6",
-    "@grapecity/wyn-integration": "^7.0.40",
-    "rxjs": "~7.4.0",
-    "tslib": "^2.3.0",
-    "webpack": "^5.65.0",
-    "webpack-dev-server": "^4.6.0",
-    "zone.js": "~0.11.4"
+    "@angular/animations": "^19.2.17",
+    "@angular/common": "^19.2.17",
+    "@angular/compiler": "^19.2.17",
+    "@angular/core": "^19.2.17",
+    "@angular/forms": "^19.2.17",
+    "@angular/platform-browser": "^19.2.17",
+    "@angular/platform-browser-dynamic": "^19.2.17",
+    "@angular/router": "^19.2.17",
+    "@wynenterprise/wyn-integration": "^9.0.286",
+    "rxjs": "^7.8.2",
+    "tslib": "^2.8.1",
+    "webpack": "^5.104.1",
+    "webpack-dev-server": "^5.2.1",
+    "zone.js": "~0.15.0"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~13.2.6",
-    "@angular/cli": "13.2.6",
-    "@angular/compiler-cli": "~13.2.6",
-    "@types/jasmine": "~3.10.0",
-    "@types/jasminewd2": "2.0.10",
-    "@types/node": "^12.11.1",
-    "codelyzer": "^6.0.0",
-    "jasmine-core": "~3.10.0",
-    "jasmine-spec-reporter": "7.0.0",
-    "karma": "~6.3.0",
-    "karma-chrome-launcher": "~3.1.0",
-    "karma-coverage": "~2.0.3",
-    "karma-jasmine": "~4.0.0",
+    "@angular-devkit/build-angular": "^19.2.19",
+    "@angular/cli": "19.2.19",
+    "@angular/compiler-cli": "^19.2.17",
+    "@types/jasmine": "~3.10.18",
+    "@types/jasminewd2": "2.0.13",
+    "@types/node": "^25.0.3",
+    "angular-eslint": "19.8.1",
+    "eslint": "^9.39.2",
+    "jasmine-core": "^5.13.0",
+    "jasmine-spec-reporter": "^7.0.0",
+    "karma": "^6.4.4",
+    "karma-chrome-launcher": "^3.2.0",
+    "karma-coverage": "~2.2.1",
+    "karma-jasmine": "^5.1.0",
     "karma-jasmine-html-reporter": "~1.7.0",
-    "protractor": "~7.0.0",
-    "ts-node": "~8.3.0",
-    "tslint": "~6.1.0",
-    "typescript": "~4.4.3"
+    "ts-node": "~8.10.2",
+    "typescript": "~5.8.0",
+    "typescript-eslint": "8.52.0"
   }
 }
