@@ -1,34 +1,37 @@
-import { Component, OnInit, Input } from '@angular/core';
+import { Component, OnInit, input, signal } from '@angular/core';
+
 
 @Component({
   selector: 'reports-list',
   templateUrl: './reports-list.component.html',
-  styleUrls: ['./reports-list.component.css']
+  styleUrls: ['./reports-list.component.css'],
+  imports: []
 })
 export class ReportsListComponent implements OnInit {
 
-  selectedReport: any;
-  reportID = "";
+  selectedReport = signal<any>(null);
+  reportID = signal("");
   rptIconSrc = "assets/report.svg";
 
-  @Input() reportSelectedCallbackFn: (reportID, title) => void;
-  @Input() token: string;
-  @Input() serverUrl: string;
-  @Input() reportsList: any;
+  reportSelectedCallbackFn = input<(reportID, title) => void>();
+  token = input<string>();
+  serverUrl = input<string>();
+  reportsList = input<any>();
 
   constructor() { }
 
-  ngOnInit(): void {    
-    
+  ngOnInit(): void {
+
   }
 
   onClick = (rpt: any) => {
     console.log(rpt);
-    this.selectedReport = rpt;
-    this.reportSelectedCallbackFn(rpt.id, rpt.name);
+    this.selectedReport.set(rpt);
+    // @ts-ignore
+    this.reportSelectedCallbackFn()(rpt.id, rpt.name);
   }
 
   getIconSrc = (rpt: any) => {
-      return this.rptIconSrc;
+    return this.rptIconSrc;
   }
 }
