@@ -1,11 +1,15 @@
 import { Component, ChangeDetectorRef, OnInit } from '@angular/core';
 import { getReportList } from './library';
-import { WynIntegration } from '@grapecity/wyn-integration';
+import { WynIntegration } from '@wynenterprise/wyn-integration';
+
+import { SignInComponent } from './sign-in/sign-in.component';
+import { ReportsListComponent } from './reports-list/reports-list.component';
 
 @Component({
-  selector: 'app-root',
-  templateUrl: './app.component.html',
-  styleUrls: ['./app.component.css']
+    selector: 'app-root',
+    templateUrl: './app.component.html',
+    styleUrls: ['./app.component.css'],
+    imports: [SignInComponent, ReportsListComponent]
 })
 export class AppComponent implements OnInit {
   loading = false;
@@ -93,7 +97,7 @@ export class AppComponent implements OnInit {
   }
 
   onSavedReport = (report) => {
-    let index = this.reportsList.findIndex(x => report.id === x.id || report.name === x.name);
+    const index = this.reportsList.findIndex(x => report.id === x.id || report.name === x.name);
     if (index === -1) {
       this.reportsList.push(report);
       this.sortReports();
